+++ assets/_src/context/provider.tsx
@@ -17,7 +17,8 @@ import SettingsContext from './settings';
  * @class
  */
 const SettingsProvider = ({ children }: { children: ReactElement[] }) => {
-	const { cfStatus, fuzion, hideSidebar, settings } = window.CFImages;
+	const { cfStatus, fuzion, hideSidebar, isNetworkAdmin, settings } =
+		window.CFImages;
 
 	const [stats, setStats] = useState(window.CFImages.stats);
 	const [modules, setModules] = useState(settings);
@@ -49,6 +50,7 @@ const SettingsProvider = ({ children }: { children: ReactElement[] }) => {
 				noticeHidden,
 				hideNotice,
 				hasFuzion,
+				isNetworkAdmin,
 				setFuzion,
 				cfConnected,
 				setCfConnected,